### 2019-07-24 #328 @florianm
Presubmission: ruODK, Client for the ODK Central API
Submitting Author: @florianm
Repository:   https://github.com/dbca-wa/ruODK/

---

```
Package: ruODK
Title: Client for the ODK Central API
Version: 0.3.1
    person(given = c("Florian", "W."),
           family = "Mayer",
           email = "Florian.Mayer@dbca.wa.gov.au",
Author: Florian W. Mayer [aut, cre]
Maintainer: Florian W. Mayer <Florian.Mayer@dbca.wa.gov.au>
Description: Utilities to access and tidy up data from ODK Central's API.
    ODK Central is OpenDataKit's clearinghouse for digitally captured data
    <https://docs.opendatakit.org/central-intro/>.
    ODK Central's API is documented at <https://odkcentral.docs.apiary.io/>.
    fs (>= 1.3.1),
    glue (>= 1.3.1),
    magrittr (>= 1.5),
    purrr (>= 0.3.2),
    readr (>= 1.3.1),
    rlang (>= 0.1.2),
    stringr (>= 1.3.1),
    tibble (>= 2.1),
    tidyr (>= 0.8.3.9000),
    xml2 (>= 1.2.0)
    covr (>= 3.0.1),
    DT (>= 0.4),
    knitr (>= 1.22),
    ggplot2 (>= 3.1.0),
    leaflet (>= 2.0.2),
    lubridate (>= 1.7.4),
    listviewer (>= 1.4.0),
    rmarkdown (>= 1.12),
Remotes: github::tidyverse/tidyr
Language: en_AU
X-schema.org-keywords: database, open-data, opendatakit, odk, api, data, dataset
X-schema.org-applicationCategory: Data Access
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

Data retrieval: `ruODK` retrieves data from ODK Central, a data clearinghouse containing data which have been digitally captured by the data collection app ODK Collect using Xforms.

As mentioned by @annakrystalli, the below categories are touched by ruODK, but don't apply:

Data extraction or munging: `ruODK` transforms and sanitises the data from ODK Central from the original format (which parses to nested lists in R) into tibbles.

Reproducibility: `ruODK` allows to script and repeat the data extraction step - the main use case it is being written for.

Geospatial data: while ODK allows to capture location data (points, lines, polygons), and `ruODK` extracts these values, `ruODK` is not primarily a spatial package.

(Scope and use cases are also mentioned in the [README](https://dbca-wa.github.io/ruODK/index.html).)

Any organisation collecting data with the [OpenDataKit](https://opendatakit.org/) suite will need to extract the data from the data clearinghouses, ODK Aggregate (outgoing) or ODK Central (new). Some may want to analyse the data straight out of ODK Central, some may need to transfer the data into another data warehouse for further post-processing, QA, and integration with other data sources.

For both use cases, `ruODK` bridges and simplifies the gap between the data sitting in ODK Central, and the data being a tibble in R, ready for further processing.

In a nutshell, `ruODK` aims to be to ODK Central what `ckanr` is to CKAN.

There are currently no other R packages wrapping the ODK Central API, neither the OData nor the RESTful API endpoints. [Relevant discussion](https://forum.opendatakit.org/t/odk-central-odata-to-r/15623/2) in the ODK forum.
Some R packages exist demonstrating basic OData access, but none worked with the OData version employed by ODK Central.
Most examples found online simply deconstruct the problem of accessing OData into parsing XML.
https://www.cdata.com/kb/tech/odata-jdbc-r.rst provides a Java-based OData client, but is not specific to ODK Central.

`ruODK` probably fulfills ROpenSci's criteria for best in class:

* `ruODK` is so far the only dedicated ODK Central API wrapper in R.
* Scope: all "data access" API endpoints, both RESTful and OData.
* Driver for development and ongoing maintenance is production use at DBCA for all Western Australian turtle monitoring projects (so far ~20k records collected through ODK since 2016) and other monitoring and research projects.
* ODK Central is in active development; DBCA is in close contact with the ODK developer community.

(TODO before ROpenSci submission: see [changelog](https://dbca-wa.github.io/ruODK/news/index.html#todo))

`ruODK` is a few functions away (seven at time of writing) from covering all "getting my data out" API endpoints, lacks some examples in the function docs, and currently uses a live ODK Central instance for testing instead of vcr/webmockr, but the existing content might be just enough to get your feedback on `ruODK`'s suitability for a later submission to ROpenSci. Any feedback is welcome!

Crosslink to [my post in ROpenSci Forum](https://discuss.ropensci.org/t/ruodk-an-api-wrapper-for-opendatakits-new-clearinghouse-odk-central/1788)

Edit: updating issue while spit-shining `ruODK`
