### 2019-07-25 ae7ffad
Add citation, clean up README
