+++ NEWS.md
@@ -21,10 +21,12 @@
 * ruODK [has a `README.Rmd`](https://devguide.ropensci.org/building.html#readme) 
   and has a [website generated by `pkgdown`](https://devguide.ropensci.org/building.html#website).
 * ruODK has documentation [generated by `roxygen2`](https://devguide.ropensci.org/building.html#documentation).
-* [Testing](https://devguide.ropensci.org/building.html#testing): test coverage
-  above 75%. 
+* [Test coverage](https://devguide.ropensci.org/building.html#testing) 100%, but
+  could use more edge cases and maybe should use `webmockr` and `vcr` instead of
+  a live ODK Central instance.
 * Spellchecks with `spelling::spell_check_package()`: added technical terms and
   function names to `inst/WORDLIST`.
+* Added citation and section in `README`.
 
 ## TODO
 * Add [examples](https://devguide.ropensci.org/building.html#examples) to docs.
