+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# ruODK: ODK Central API wrapper <img src="man/figures/ruODK.png" align="right" alt="Are you ODK?" width="120" />
+# ruODK: Client for the ODK Central API <img src="man/figures/ruODK.png" align="right" alt="Are you ODK?" width="120" />
 
 <!-- badges: start -->
 
@@ -36,34 +36,68 @@ provides access to its data and functionality through both an OData and
 a RESTful API with a comprehensive and interactive
 [documentation](https://odkcentral.docs.apiary.io/#reference/odata-endpoints).
 
-ruODK aims:
+`ruODK`’s scope:
 
-  - To wrap all ODK Central API endpoints with a focus on data access.
+  - To wrap all ODK Central API endpoints with a focus on **data
+    access**.
   - To provide working examples of interacting with the ODK Central API.
   - To provide convenience helpers for the day to day tasks when working
-    with ODK Central data in R.
+    with ODK Central data in R: **data munging** the ODK Central API
+    output into tidy R formats.
+
+`ruODK`’s use cases:
+
+  - Smaller projects:
+      - Data collection: ODK Collect `%>%`
+      - Data clearinghouse: ODK Central `%>%`
+      - Data analysis and reporting: `Rmd` (ruODK) `%>%`
+      - Publishing and dissemination:
+        [`ckanr`](https://docs.ropensci.org/ckanr/),
+        [`CKAN`](https://ckan.org/)
+  - Larger projects:
+      - Data collection: ODK Collect `%>%`
+      - Data clearinghouse: ODK Central `%>%`
+      - ETL pipeline into data warehouses: `Rmd` (ruODK) `%>%`
+      - QA: in data warehouse `%>%`
+      - Reporting: `Rmd` `%>%`
+      - Publishing and dissemination:
+        [`ckanr`](https://docs.ropensci.org/ckanr/),
+        [`CKAN`](https://ckan.org/)
 
 Out of scope:
 
-  - Wrap “management” API endpoints. The ODK Central GUI is a fantastic
-    interface for the management of users, roles, permissions, projects,
-    and forms.
-  - Extensive data visualisation. We show only minimal examples of data
-    visualisation and presentation, mainly to illustrate the example
-    data.
+  - To wrap “management” API endpoints. The ODK Central GUI already is a
+    fantastic interface for the management of users, roles, permissions,
+    projects, and forms. In fact, it is a [VueJS
+    application](https://github.com/opendatakit/central-frontend/)
+    working on the “management” API endpoints of the ODK Central
+    backend.
+  - To provide extensive data visualisation. We show only minimal
+    examples of data visualisation and presentation, mainly to
+    illustrate the example data. Once the data is in your hands as tidy
+    tibbles… u r ODK\!
 
 ## Install
 
 You can install ruODK from GitHub with:
 
 ``` r
-# install.packages("remotes")
-remotes::install_github("dbca-wa/ruODK")
+# install.packages("devtools")
+remotes::install_github("dbca-wa/ruODK", dependencies = TRUE)
 ```
 
 ## ODK Central
 
-First, we need some data to play with\!
+### ODK Central instance
+
+First, we need an ODK Central instance and some data to play with\!
+
[…]
