### 2019-07-25 462ebb5
Close #10: rename OData functions
