+++ man/prepend_uuid.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_attachment.R
+% Please edit documentation in R/attachment_get.R
 \name{prepend_uuid}
 \alias{prepend_uuid}
 \title{Prepend a leading "uuid:" to any string, e.g. an md5 hash.}

[…]
