### 2019-07-25 f6d4a54
Expand test coverage for odata_submissions_get and add examples
