+++ R/odata_submissions_get.R
@@ -1,25 +1,36 @@
 #' Retrieve /Submissions from an OData URL ending in .svc as list of lists
 #'
+#'
+#' @details See https://odkcentral.docs.apiary.io/#reference/odata-endpoints/odata-form-service/data-document
 #' @template param-pid
 #' @template param-fid
-#' @param sub The submission EntityType name,
-#'            default: "Submissions" (the main form).
+#' @param table The submission EntityType, or in plain words, the table name.
+#'            Default: "Submissions" (the main table).
 #'            Change to "Submissions.GROUP_NAME" for repeating form groups.
 #'            The group name can be found in the metadata under
 #'            DataServices.Schema.[EntityType].Name
+#' @param skip The number of rows to be omitted from the results.
+#'             Example: 10, default: NA (none skipped).
+#' @param top The number of rows to return.
+#'            Example: 100, default: NA (all returned).
+#' @param count If TRUE, an `@odata.count` property will be returned in the
+#'              response from ODK Central. Default: FALSE.
+#' @param wkt If TRUE, geospatial data will be returned as WKT (Well Known Text)
+#'            strings. Default: FALSE, returns GeoJSON structures.
 #' @template param-auth
 #' @importFrom glue glue
 #' @importFrom httr add_headers authenticate content GET
 #' @export
-#' @return A list of two named lists, value and context. `value` contains
-#'     the submissions, which can be "rectangled" using
+#' @return A nested list of lists.
+#'         `value` contains the submissions, which can be "rectangled" using
 #'     `tidyr::unnest_wider("element_name")`.
-#'     `context` is the URL of the metadata.
-#'     See spec at
+#'     `@odata.context` is the URL of the metadata.
+#'     `@odata.count` is the total number of rows in the table.
+#'     See spec at https://odkcentral.docs.apiary.io/#reference/odata-endpoints/odata-form-service
 #' @examples
 #' \dontrun{
 #'
-#' Replace with your own working url, pid, fid, credentials:
+#' # Replace with your own working url, pid, fid, credentials:
 #' pid <- 14
 #' fid <- "build_Flora-Quadrat-0-2_1558575936"
 #' url <- "https://sandbox.central.opendatakit.org"
@@ -27,11 +38,32 @@
 #' pw <- Sys.getenv("ODKC_TEST_PW")
 #'
 #' # With default credentials, see vignette("setup", package = "ruODK")
-#' data <- odata_submissions_get(pid, fid, sub = "Submissions")
+#' data <- odata_submissions_get(pid, fid, table = "Submissions")
 #'
 #' # With explicitly set credentials
-#' data <- odata_submissions_get(pid, fid, sub = "Submissions",
-#'                               url = url, un = un, pw = pw)
+#' data <- odata_submissions_get(pid, fid,
+#'   table = "Submissions",
+#'   url = url, un = un, pw = pw
+#' )
+#'
+#' # Skip one row, return the next 1 rows (top), include total row count
+#' data <- odata_submissions_get(pid, fid,
+#'   table = "Submissions",
+#'   skip = 1, top = 1, count = TRUE,
+#'   url = url, un = un, pw = pw
+#' )
+#'
+#' # Parse point coordinates into lat, lon, alt
+#' data <- odata_submissions_get(pid, fid,
+#'   table = "Submissions", wkt = FALSE,
+#'   url = url, un = un, pw = pw
+#' )
+#'
+#' # Parse point coordinates into WKT like "POINT (115.8840312 -31.9961844 0)"
+#' data <- odata_submissions_get(pid, fid,
+#'   table = "Submissions", wkt = TRUE,
+#'   url = url, un = un, pw = pw
+#' )
 #'
[…]
