### 2019-07-25 d2875a0
Update OData vignette: spatial as WKT or GeoJSON
