### 2019-07-25 726b3b6
snake_case columns in project_list and project_detail
