### 2019-07-25 #328 @annakrystalli
Hi @florianm 👋and thanks for your presubmission enquiry,

I'd not heard of this framework before but it looks really cool! I listened to the podcast @sckott posted in response to your forum post last night, will definitely be checking it out.

In terms of the topic, you've ticked data retrieval, extraction and reproducibility but we feel the correct category is data retrieval only. I appreciate that it still needs some work but in principle, we feel we can invite a full submission when ready.
