### 2019-07-25 fcf33b9
Close #8: first cut of paper.md for JOSS
