+++ inst/joss/paper.bib
@@ -1,0 +1,58 @@
+@Misc{ruodk,
+  title = {ruODK: Client for the ODK Central API},
+  author = {Florian W. Mayer},
+  note = {R package version 0.3.1},
+  year = {2019},
+  url = {https://github.com/dbca-wa/ruODK},
+}
+
+@online{github,
+  author = {GitHub},
+  title = {GitHub.com},
+  year = 2008,
+  url = {https://github.com},
+  urldate = {2019-07-25}
+}
+
+@online{odk,
+  author = {OpenDataKit},
+  title = {opendatakit.org},
+  year = 2019,
+  url = {https://opendatakit.org/},
+  urldate = {2019-07-25}
+}
+
+@online{odkdocs,
+  author = {OpenDataKit Documentation},
+  title = {docs.opendatakit.org},
+  year = 2019,
+  url = {https://docs.opendatakit.org/},
+  urldate = {2019-07-25}
+}
+
+@online{odkapi,
+  author = {OpenDataKit API},
+  title = {odkcentral.docs.apiary.io},
+  year = 2019,
+  url = {https://odkcentral.docs.apiary.io/},
+  urldate = {2019-07-25}
+}
+
+@inproceedings{hartung,
+ author = {Hartung, Carl and Lerer, Adam and Anokwa, Yaw and Tseng, Clint and Brunette, Waylon and Borriello, Gaetano},
+ title = {Open Data Kit: Tools to Build Information Services for Developing Regions},
+ booktitle = {Proceedings of the 4th ACM/IEEE International Conference on Information and Communication Technologies and Development},
+ series = {ICTD '10},
+ year = {2010},
+ isbn = {978-1-4503-0787-1},
+ location = {London, United Kingdom},
+ pages = {18:1--18:12},
+ articleno = {18},
+ numpages = {12},
+ url = {http://doi.acm.org/10.1145/2369220.2369236},
+ doi = {10.1145/2369220.2369236},
+ acmid = {2369236},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {ICTD, client-server distributed systems, mobile computing, mobile phones},
+}
