### 2019-07-25 #328 @florianm
Cheers @annakrystalli!
ODK is indeed the bees knees.

The categories are updated now, and I've included a paper to JOSS (is `/inst/joss` a good place for the paper.md?)
