### 2019-07-25 #328 @annakrystalli
Great!

I think just the root of the repo is fine as it doesn't really need to be included in the package when installed (it will stand as it's one publication in JOSS).  It will be easier for reviewers to find too. In this case it would also need to be included in `.Rbuildignore`

Having said that, we don't have specific guidelines on the location of `paper.md` so it certainly wouldn't be a problem to store it in `inst/joss`. Just make sure reviewers know where it is. :)
