### 2019-07-25 #328 @florianm
Cool - didn't want to clutter root with non-standard files, but the paper can go in either place. If I add the remaining functions (with docs, examples, tests, and examples in README and in vignette), is there any other glaring error or omission I could fix or add prior to submission?
